출근 시간 퇴근 시간 계산으로 수정

+++ 미니프로젝트/출퇴근 사내 시스템/miniproject/src/main/java/chuchu/miniproject/dto/worker/response/ResponseGetWorkList.java
@@ -6,6 +6,6 @@ public record ResponseGetWorkList(
 
         List<ResponseListGetWorkList> detail,
 
-        Integer sum
+        Long sum
 ) {
 }
