+++ 미니프로젝트/출퇴근 사내 시스템/miniproject/src/main/java/chuchu/miniproject/dto/worker/response/ResponseListGetWorkList.java
@@ -5,6 +5,7 @@ import java.time.LocalDate;
 public record ResponseListGetWorkList(
         LocalDate date,
 
-        Integer workingMinute
+        Long workingMinute
+
 ) {
 }
